requirements.txt
streamlit==1.37.1
python-dotenv==1.0.1
google-genai==0.7.0
pandas==2.2.2
numpy==2.0.1
matplotlib==3.9.1
plotly==5.22.0

